Correct issues with manipulating the interrupt mask that can result in an interrupt storm.  This could only happen in station mode with devices that support h/w phy counters.
There were two issues. sc_imask is intended to hold the set of
interrupts the driver is prepared to field.  One must always clear
an interrupt in the hardware before removing it from this mask,
otherwise an interrupt might be delivered and then discarded.  This
can result in an interrupt storm for interrupts such as the MIB
interrupt which is repeatedly posted until the MIB counters are
cleared.  Resolve this by always clearing interrupts from the IMR
before sc_imask.

The second issue is that setting the interrupt state with ath_hal_intrset
is prone to races because multiple registers must be manipulated
for 5211 and 5212 mac's.  Previously, when a MIB interrupt was
fielded, the driver would clear the interrupt and activate a tasklet
to process the event (clearing the MIB registers and re-enabling
the MIB interrupt).  However that setup two concurrent threads that
could update the interrupt mask and so opened a race.  To resolve
this we eliminate the mib tasklet and process a MIB interrupt
immediately so the only possible concurrency is with re-entry to
the interrupt handler.

Submitted by:	Dale Whitfield