HAL update by Sam Leffler: HAL version 0.9.16.3
Attached are the changes to use the 0.9.16.3 hal I released a while 
back.  This also nukes the CTS bursting update stuff though the hal 
release still has it.

In addition there are some revisions to the tx99 support that is not 
being released; you can take it if you want;

Past that is some lightly-tested support for ahdemo mode.  Looks like I 
left out the obvious change to wlanconfig.