Fix a "cosmetic" problem with the quater dBm value parsed from the HAL to the IEEE 802.11 stack.
The HAL delivers the values for min and max power in .25 dBm values
already, the values are parsed in to the priv struct converting the
values to .25 dBm again. So effectively ichan->ic_maxpower and
ichan->ic_minpower are in .125 dBm values!

These values don't influence the Tx power used by the card, but are
used in information ellements like the 802.11h information ellement.

Signed-off-by: R. van Brenk <r.vanbrenk@hopling.com>