Use two stage compile with hal.o replacement to fool Linux 2.6.17-rc1
It's getting increasingly hard to fool the kernel build system into
thinking that hal.o has sources.  This patch makes compiling the module
a two stage process.

Initially, a dummy hal.c is created to pacify the build system. 
ath_hal.ko is compiled using it.

Then the real hal.o (disguised as hal.obj) is copied on top of the dummy
hal.o and touched.  ath_hal.ko and temporary ath_hal.* files are
removed.  The second stage compile is run, and this time it picks up the
right hal.o.