1) Backout the wext 19 changes to ieee80211_init that caused this    problem (r1499).
2) Fix the real bug that made me change ieee80211_init in the first
   place.

The real bug was a weird one. wpa_supplicant was setting the
authentication algorithm to "open system", which madwifi was
interpreting as "no encryption." However, those apparently do not
mean the same thing. The original "fix" was to avoid rescanning
after wpa_supplicant had set the authentication algorithm, which
allowed the association to complete.

The new fix is to ignore the authentication algorithm specified
by wpa_supplicant. wpa_supplicant falls back to setting the same
information via the IW_ENCODE flags, so the authentication system
still works.

Signed-Off-By: Brian Eaton <eaton.lists@gmail.com>