Full rewrite of get_arch_target.sh
This was destined for Makefile.inc originally, and the code is still in
a format that would allow easy integration into a makefile.

If .config is missing, make it an error.

Separate ARCH and TARGET calculation.  ARCH can often be determined much
more reliably, so that part of code won't need to change much.

CONFIG_SPARC is not defined for 32-bit SPARC.  Use CONFIG_SPARC32
instead.

Determine endianess more reliably.  Don't use variable endianess for
CPUs when there is no choice in the kernel or in the HAL.

Recognize CONFIG_CPU_SA110 and CONFIG_CPU_SA1100 as xscale.

Initial attempt at distinguishing misp1, mips and mipsisa32 targets.