The average RSSI calculation (ieee80211_getrssi in ieee80211_node.c) does not correctly calculate the average RSSI when operating in adhocdemo mode.
The problem seems to occur because the node table contains an entry
for the local interface as well as any remote nodes that are
associated. This local entry often has very high or very low RSSI
readings.

The attached patch fixes this problem be excluding the local
interface from inclusion in the average. This behaviour is similar to
what occurs in ieee80211_linux.c when the associated_sta file is
being generated.

Signed-off-by: Matt Brown <matt@mattb.net.nz>

Closes #685