This patch adds support for reporting statistics on the operation of the rate control modules via /proc. Currently only the sample rate module actually provides any statistics.
The statistics reported for the sample module were present in the
madwifi-old module but did not seem to make it into madwifi-ng. I
have made some minor changes to how these statistics are reported:

 1. Statistics are now reported via /proc entries instead of using
    sysctls. Sysctls are intended as a method for the user to
    configure parameters of the module and should not be misued for
    reporting statistics.
 2. A new symbol ath_rate_dynamic_proc_register is exported from each
    rate control module. This function is called by the driver just
    after a new VAP has been created to allow the rate control module
    to register the appropriate proc entries. As described in ticket
    #728, all further management of these entries will be handled by
    the VAP, which will call the functions supplied by the rate
    control module to retrieve data when necessary.

The information that this patch adds back into the driver is very
useful and is referred to in the documentation which currently causes
some confusion when the statistics is not found. The documentation
should be updated to reflect that the statistics will now live at
/proc/net/madwifi/<ifname>/ratestats_{250,1600,3000}.

Closes #729

Signed-off-by: Matt Brown <matt@mattb.net.nz>