Use _shipped objects instead of faking dependencies for Linux 2.6
Another attempt at doing The Right Thing with pre-compiled HAL.  Linux
2.6 allows to suppress building of any file if there is a file with the
same name plus "_shipped" at the end.  This is as close as it can get to
the situation we have with HAL.

So, instead of creating a fake .cmd file with empty dependencies, the
_shipped mechanism is now used for 2.6 kernels.  2.4 kernels are not
affected, except that there will be no useless .cmd files lying around.