Update, remove references to TOOLPREFIX
Mention that the kernel source package is not necessary.  Be more
specific about package names of the kernel development packages.

Don't suggest overriding KERNELRELEASE.  Generally, try not to advice
users to override variables that Madwifi can discover on its own.

HAL *.inc files are in hal/pub, not in hal/linux.  Don't list them all.

Mention that incompatible flags may mean wrong target.

Don't quote the whole *.inc file.  Absolute majority of users have no
need to read and understand it to install Madwifi.

List modules only once, without suffixes and directories.  The list is
the same for Linux 2.4 and 2.6.  Adjust the list to reflect the current
sources.

Remove references to the card manager.  We don't have a PCMCIA driver! 
Only one module, ath_pci, needs to be loaded manually (an only if the
system doesn't load it on startup already).