Use a "custom" rule for 2.6 kernels
Change the build system to conform to the "Custom kbuild commands"
section in the kernel's Documentation/kbuild/makefiles.txt

Note that we don't need FORCE since the uudecode flags are not user
configurable.  We don't need to add the target to $(targets) because
it's already there (as a dependency of the module).