Revert removal of TARGET calculation in r1726
TARGET still referenced and needed to process opt_ah.h correctly, which
is needed to test for AH_NEED_DESC_SWAP in if_ath.c.

This creates the third place where TARGET is calculated, and of course
if needs refactoring as soon as possible, but the objective now is to
revert the damage by the apparenly untested changes made to this file.