This patch fixes a bug which caused invalid sequence numbers appear in packets originating from the device when local monitor mode is used (that is: monitor as one of VAPs).
It happened because sequence numbers are assigned by the hardware and can be
reclaimed after processing tx descriptor.

The patch introduces setting the sequence numbers of captured frames to the
ones assigned by the hardware. It improves local monitor mode usability a
bit, however it is still severly broken because of other issues (out of scope
of this patch).

Signed-off-by: Michal Wrobel <xmxwx@asn.pl>