Remove special treatment of AHB bus.
All special treatment of AHB bus is just a bunch of obsolete hacks.

TOOLPATH works around some cross-compiler issues not specific to AHB
bus.  It may be useful, but it shouldn't be required.  INCS can be
defined on the command line just fine.

TARGET being mipsisa32-be-elf is likely incorrect.  There are new HAL
binaries for system-on-chip systems, so TARGET needs to be overridden
anyway.

KERNELPATH can be guessed just the way it's done for PCI builds.  There
is no reason why it won't work for AHB bus.

KMODPATH hardcodes a long path that would be wrong in general case.  The
default used by the PCI build is much saner.

PCI checks should check BUS to be PCI for readability if nothing else.