In a multi-vap scenario (for example, a STA or WDS-REPEATER vap on the same device as an AP vap) an IEEE80211_MLME_DISASSOC/DEAUTH (for example, from hostapd upon startup) will call ieee80211_node_leave for all nodes in the node table, regardless of what vap they are associated with.
Instead, ieee80211_node_leave should only be called for nodes that are on the
specified vap (Note that this same mis-behavior of iterating over all nodes
regardless of vap may be alsmo present elsewhere in madwifi)

This can cause a number of issues, including driver crash when nodes get
dropped that shouldn't be.

See #969 for details about how to reliably reproduce this problem.

Signed-off-by: Tim Harvey <tim_harvey@yahoo.com>