Define bf_flags field without ATH_SUPERG_FF enabled
bf_flags is only used to find out whether ACK was expected for the Tx
frame.  This is used to decide whether to process the frame by the rate
control algorithm.  There is nothing SuperG specific in that field.

Reorder the fields to allow a denser structure with or without
ATH_SUPERG_FF.