Remove references to CONFIG_SYSCTL and CONFIG_NET_WIRELESS
CONFIG_SYSCTL is required, and adding support for configurations without
sysctl would be a waste of time at this point.

Wireless extensions are on the way out, but the references to
CONFIG_NET_WIRELESS are in dead code anyway, so just clean them up as
the same is done in the trunk.