Include compat.h from the sources, not from the command line
Including it through -include leads to a compile error on x86_64, which
is hard to fix reliably and without risk of breakage with future kernels
and glibc.

When compat.h is the first to be included, __le64 is defined in such way
that u_int64_t is redefined when linux/types.h is included by one of the
glibc headers.

If compat.h is included after the system headers, it has a chance to
provide only necessary definitions.

Besides, wpa_supplicant and hostapd include compat.h, so it's better to
stay compatible with the code we cannot control.