Remove all unnecessary variable initializations
The compiler is quite good at finding uninitialized variables. 
Initializing a variable with a bogus or a "reasonable but untested"
value creates risk that it would be used following some code changes,
breaking the algorithm in a subtle way.

It's much better to have a warning at the compile time, so that the
changes can be examined as soon as they have been made.