Fix splitting MAC addresses on big-endian systems
Bytewise copying from a MAC address to an integer is incorrect since it
assumes little endian byte order.  Make it explicit that the MAC address
is treated as two little-endian values 32-bit and 16-bit long
respectively.  Convert the data from little endian to the host order,
since the arguments to AR5K_REG_WRITE are in the host order.