Use safer and saner array initializers for wireless handlers
Instead of relying on the correct order of members in ieee80211_handlers
and ieee80211_priv_handlers initializers, specify indices explicitly. 
Eliminate holes in the tables, as they are created by the compiler now.

This change allows the compiler to check that the corresponding handler
numbers are defined and not duplicated.  In the case of
ieee80211_priv_handlers, the names were nowhere near the initializer,
which made it hard to understand the code.

Remove ieee80211_ioctl_getscanresults() from ieee80211_priv_handlers.
It has no corresponding private handler and it's a generic ioctl
handler, not an wireless private handler.