I found two major impediments to using athdebug and 80211debug:
1. When no device exists, using 80211debug or athdebug without arguments will cause an error to be displayed, rather than usage information.

2. Flags are entered and displayed in either hexadecimal (which is efficient but non-friendly) or symbolic names. However, symbolic names are never displayed and are only available if you have man pages installed (not common on access points).

The attached patch corrects both situations. For #1 we add usage message if this occurs. For #2 we add a tabular mapping of symbolic names to hexadecimal values in the usage information and after each invocation we display the same table with an indicator of which flags are enabled.

These changes greatly improve the usability of the debug flags.