Remove workaround for pre-releases of Fedora 7
It was just a matter of convenience before Fedora 7 release, but now it's
dead code.  Fedora 7 was released with Linux 2.6.21, so the fix is
irrelevant to it.

A nice side effect is that linux/sysctl.h doesn't need to be included
from compat.h any longer, which makes compat.h less intrusive.

It also eliminates some sparse warnings, even though they are likely
caused by a sparse bug.