This patch fixes a bug in the usage of wireless extensions to pass structures via driver-specific extensions (iwpriv).
The structures were being identified to wext as an array of bytes, which was ok
as long as the byte count fit within the bit field (see IW_PRIV_SIZE_MASK).  If
the size exceeded the bit field, we could get a buffer overflow.

WEXT looks at the size of the element and the number of elements and allocates a
buffer large enough to hold the data.  It handles copying the data to/from the
user's provided buffer and provides an "extra" pointer to the handler.

The poblem would occur when the request lists a buffer length greater than this
maximum number of bytes (again, see IW_PRIV_SIZE_MASK) but the actual buffer
provided by wext was smaller.

This patch provides macros for constructing size and type encodings for structs
to be exposed in this manner to wireless extensions.

It prefers to list structures as byte arrays whenever possible, but degenerates
to 4-byte integer arrays when necessary in order to correctly encode the buffer
space needed.