Guard against zero rates
Using a rate of 0 in the txdescriptor will cause the hardware to never stop
sending "noise" so it will block the channel for all other nodes. Thus it is
important to avoid this situation.

In case we receive bogus frames which contain rates of 0 we silently delete
them in ieee80211_fix_rate().

Note that this does not contain the check for multi-rate retries as in the
original changeset, as the rate API changed and the extended descriptor is set
up by the rate modules.

Reference changesets:
* http://madwifi.org/changeset/3013
* http://madwifi.org/changeset/3015
* http://madwifi.org/changeset/3035