Fix for bridging.  We modify the packet's contents.  Linux bridging code assumes it can clone the packet and send it to multiple interfaces without having it changed out from under it.  Therefore, we need to clone each skb prior to modifying the copy.  This fixes some very nasty bugs with bridged VAPs (esp with multiple VAPs in a bridge and multicast/broadcast traffic).
The symptom would be fragments of 802.11 header showing up in the ethernet header because of skb header offsets being corrupted by MadWifi devices on lower ports which then get forwarded as-is to wired ports.

Sorry for the lame explanation, but the short end of the story is that you have to copy the skb before you modify it if you are going to play nice with bridging.