Ensure non-AP STA VAPs correctly notify the AP when when their power save state changes.
Previously, a confusion between the node's power save state and the VAP's
power save state was causing all nulldata frames to be sent with 0 in the
PwrMgt bit.  This was causing the AP to not notice when the STA went off
channel to scan and hence not buffer frames.

We now update the VAP's power save state so that nulldata frames are sent with
the correct bits set. We also clean up some logic surrounding the use of XOR.

Reference changesets:
* http://madwifi.org/changeset/2330
* http://madwifi.org/changeset/2345
* http://madwifi.org/changeset/3062
* http://madwifi.org/changeset/3065