Avoid flexible arrays in structures for gcc-2.95 compatibility
Use zero-length arrays instead.  Flexible arrays are ISO C99, and
zero-length arrays is an older gcc extension.  But apparently we still
have users with gcc-2.95, and it's the only issue preventing MadWifi
compilation, so it would be better to fix it for now.  We could switch
back to C99 flexible arrays once Linux 2.4 support is dropped.