The RX tasklet must process the CABQ when it is setup and non-empty, even if the HAL says it is not "active".
HAL will say the CABQ is not active if the trigger has fired and the CABQ has been serviced.  Therefore, the conditions under which we were skipping CABQ are EXACTLY when we need to check it.

This was leading to large numbers of rx buffers consumed in the CABQ (multicast, etc) and thus the RX queue would deteriorate over time until RX overruns would eventually start to be a big problem.

I also include better diagnostics.